requirements.txt
langgraph>=0.0.36
langchain-core>=0.1.0
langchain-openai>=0.0.1
typing-extensions>=4.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0